requirements.txt
Flask==2.3.3
PyYAML==6.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
requests==2.31.0 